ref(feedback): De-couple button from modal
De-couple the feedback button from the modal so that we can show the modal using other components as the actor.

This change does the following:

* Move the form in `<FeedbackModal>` to `<FeedbackForm>`
* Move business logic of the modal/form into `<FeedbackModal>` (from `<FeedbackWidget>`)
* `<FeedbackModal>` requires `children` as a render function
* `<FeedbackWidget>` is refactored to use new `<FeedbackModal>` and remains the same (shows our `<FeedbackButton>` which controls the modal)
