+++ src/components/feedback/feedbackWidget.tsx
@@ -1,74 +1,24 @@
-import React, {Fragment, useEffect, useState} from 'react';
+import React from 'react';
 
 import {FeedbackButton} from './feedbackButton';
 import {FeedbackModal} from './feedbackModal';
-import {FeedbackSuccessMessage} from './feedbackSuccessMessage';
-import {sendFeedbackRequest} from './sendFeedbackRequest';
 
-interface FeedbackForm {
-  comment: string;
-  email: string;
-  name: string;
+interface FeedbackWidgetProps {
+  title?: string;
 }
 
-async function sendFeedback(
-  data: FeedbackForm,
-  replayId: string,
-  pageUrl: string
-): Promise<Response | null> {
-  const feedback = {
-    message: data.comment,
-    email: data.email,
-    replay_id: replayId,
-    url: pageUrl,
-  };
-  const response = await sendFeedbackRequest(feedback);
-  return response;
-}
-
-export default function FeedbackWidget() {
-  const [open, setOpen] = useState(false);
-  const [showSuccessMessage, setShowSuccessMessage] = useState(false);
-
-  useEffect(() => {
-    if (!showSuccessMessage) {
-      return () => {};
-    }
-    const timeout = setTimeout(() => {
-      setShowSuccessMessage(false);
-    }, 6000);
-    return () => {
-      clearTimeout(timeout);
-    };
-  }, [showSuccessMessage]);
-
-  const handleSubmit = (data: FeedbackForm) => {
-    const replay = window.Sentry?.getCurrentHub()
-      ?.getClient()
-      ?.getIntegration(window.Sentry?.Replay);
-
-    // Prepare session replay
-    replay?.flush();
-    const replayId = replay?.getReplayId();
-
-    const pageUrl = document.location.href;
-
-    sendFeedback(data, replayId, pageUrl).then(response => {
-      if (response) {
-        setOpen(false);
-        setShowSuccessMessage(true);
-      } else {
-        // eslint-disable-next-line no-alert
-        alert('Error submitting your feedback. Please try again');
-      }
-    });
-  };
+/**
+ * The "Widget" connects the default Feedback button with the Feedback Modal
+ */
+export default function FeedbackWidget({title = 'Got Feedback?'}: FeedbackWidgetProps) {
+  // Don't render anything if Sentry is not already loaded
+  if (!window.Sentry?.getCurrentHub?.()) {
+    return null;
+  }
 
   return (
-    <Fragment>
-      {!open && <FeedbackButton onClick={() => setOpen(true)} />}
-      <FeedbackModal open={open} onSubmit={handleSubmit} onClose={() => setOpen(false)} />
-      <FeedbackSuccessMessage show={showSuccessMessage} />
-    </Fragment>
+    <FeedbackModal title={title}>
+      {({open, showModal}) => (open ? null : <FeedbackButton onClick={showModal} />)}
+    </FeedbackModal>
   );
 }
